Use Photon for upstream thumbnail compression
