+++ api/catalog/api/views/media_views.py
@@ -1,28 +1,17 @@
 import logging
-from urllib.parse import urlencode, urlparse
 
-from django.conf import settings
-from django.core.cache import cache
-from django.http.response import HttpResponse
 from rest_framework import status
 from rest_framework.decorators import action
 from rest_framework.exceptions import APIException
 from rest_framework.response import Response
 from rest_framework.viewsets import ReadOnlyModelViewSet
 
-import requests
-from sentry_sdk import capture_exception
-
 from catalog.api.controllers import search_controller
 from catalog.api.models import ContentProvider
 from catalog.api.serializers.provider_serializers import ProviderSerializer
+from catalog.api.utils import photon
 from catalog.api.utils.pagination import StandardPagination
 from catalog.custom_auto_schema import CustomAutoSchema
-
-
-class UpstreamThumbnailException(APIException):
-    status_code = status.HTTP_424_FAILED_DEPENDENCY
-    default_detail = "Could not render thumbnail due to upstream provider error."
 
 
 parent_logger = logging.getLogger(__name__)
@@ -167,7 +156,7 @@ class MediaViewSet(ReadOnlyModelViewSet):
     def thumbnail(self, image_url, request, *_, **__):
         serializer = self.get_serializer(data=request.query_params)
         serializer.is_valid(raise_exception=True)
-        return self._get_proxied_image(
+        return photon.get(
             image_url,
             accept_header=request.headers.get("Accept", "image/*"),
             **serializer.validated_data,
@@ -190,101 +179,3 @@ class MediaViewSet(ReadOnlyModelViewSet):
         else:
             ip = request.META.get("REMOTE_ADDR")
         return ip
-
-    THUMBNAIL_PROXY_COMM_HEADERS = {
-        "User-Agent": settings.OUTBOUND_USER_AGENT_TEMPLATE.format(
-            purpose="ThumbnailGeneration"
-        )
-    }
-
-    @staticmethod
-    def _thumbnail_proxy_comm(
-        path: str,
-        params: dict,
-        headers: tuple[tuple[str, str]] = (),
-    ) -> tuple[requests.Response, int, str]:
-        logger = parent_logger.getChild("_thumbnail_proxy_comm")
-        proxy_url = settings.THUMBNAIL_PROXY_URL
-        query_string = urlencode(params)
-        upstream_url = f"{proxy_url}/{path}?{query_string}"
-        logger.debug(f"Image proxy upstream URL: {upstream_url}")
-
-        try:
-            compiled_headers = MediaViewSet.THUMBNAIL_PROXY_COMM_HEADERS | {
-                k: v for k, v in headers
-            }
-            upstream_response = requests.get(
-                upstream_url, timeout=10, headers=compiled_headers
-            )
-
-            res_status = upstream_response.status_code
-            content_type = upstream_response.headers.get("Content-Type")
-            logger.debug(
-                "Image proxy response "
-                f"status: {res_status}, content-type: {content_type}"
-            )
-
-            return upstream_response, res_status, content_type
-        except requests.ReadTimeout as exc:
-            # Count the incident so that we can identify providers with most timeouts.
-            domain = urlparse(params["url"]).netloc
-            key = f"{settings.THUMBNAIL_TIMEOUT_PREFIX}{domain}"
-            try:
-                cache.incr(key)
-            except ValueError:  # Key does not exist.
-                cache.set(key, 1)
-
-            capture_exception(exc)
-            raise UpstreamThumbnailException(
-                f"Failed to render thumbnail due to timeout: {exc}"
-            )
-        except requests.RequestException as exc:
-            capture_exception(exc)
-            raise UpstreamThumbnailException(f"Failed to render thumbnail: {exc}")
-        except Exception as exc:
-            capture_exception(exc)
-            raise UpstreamThumbnailException(
-                f"Failed to render thumbnail due to unidentified exception: {exc}"
-            )
-
-    @staticmethod
-    def _get_proxied_image(
-        image_url: str,
-        accept_header: str = "image/*",
-        is_full_size: bool = False,
-        is_compressed: bool = True,
-    ):
-        width = settings.THUMBNAIL_WIDTH_PX
-        if is_full_size:
-            info_res, *_ = MediaViewSet._thumbnail_proxy_comm(
-                "info", {"url": image_url}
-            )
-            info = info_res.json()
-            width = info["width"]
-
-        params = {
-            "url": image_url,
-            "width": width,
-        }
-
-        if is_compressed:
-            params |= {
-                "quality": settings.THUMBNAIL_JPG_QUALITY,
-                "compression": settings.THUMBNAIL_PNG_COMPRESSION,
-            }
-        else:
-            params |= {
-                "quality": 100,
-                "compression": 0,
-            }
-
-        if "webp" in accept_header:
-            params["type"] = "auto"  # Use ``Accept`` header to determine output type.
-
-        img_res, res_status, content_type = MediaViewSet._thumbnail_proxy_comm(
-            "resize", params, (("Accept", accept_header),)
-        )
-        response = HttpResponse(
-            img_res.content, status=res_status, content_type=content_type
-        )
-        return response
